src/python/requirements.txt
typer>=0.9.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
boto3>=1.28.0,<2.0.0
azure-identity>=1.13.0,<2.0.0
azure-mgmt-network>=23.0.0,<24.0.0
azure-mgmt-compute>=30.0.0,<31.0.0
azure-mgmt-resource>=23.0.0,<24.0.0
google-cloud-asset>=3.20.0,<4.0.0
jinja2>=3.1.0,<4.0.0
rich>=13.0.0,<14.0.0
requests>=2.31.0,<3.0.0
netaddr>=0.9.0,<1.0.0
